refactor: isolate ESM2 position id routing
